+++ python/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the AIStore Python SDK project are documented in this fil
 We structure this changelog in accordance with [Keep a Changelog](https://keepachangelog.com/) guidelines, and this project follows [Semantic Versioning](https://semver.org/).
 
 ---
+## Unreleased
+
+### Added
+
+- Add `ETLRuntimeSpec` class to formalize runtime configuration.
+- Add `etl.init(image, command, …)` for simplified setup using only image and command.
+
 ## [1.14.0] - 2025-05-27
 
 ### Added